SPDBT-3378
