+++ src/Spd.Presentation.Screening/ClientApp/src/app/shared/components/file-upload.component.ts
@@ -6,72 +6,7 @@ import { HotToastService } from '@ngxpert/hot-toast';
 import { SPD_CONSTANTS } from 'src/app/core/constants/constants';
 import { UtilService } from 'src/app/core/services/util.service';
 import { DialogComponent, DialogOptions } from './dialog.component';
-
-export enum DocumentTypeCode {
-	Image = 'IMAGE',
-	Pdf = 'PDF',
-	Word = 'WORD',
-}
-
-export interface IconType {
-	icon: string;
-	label: string;
-	color?: string;
-}
-
-export class FileUploadHelper {
-	private static _FILE_ICONS: Record<DocumentTypeCode, IconType> = {
-		PDF: { icon: 'picture_as_pdf', label: 'PDF' },
-		WORD: { icon: 'article', label: 'Microsoft Word' },
-		IMAGE: { icon: 'image', label: 'Image file' },
-	};
-
-	private static _MIME_TYPE_DOCUMENT_TYPE_MAP: Record<string, DocumentTypeCode> = {
-		'application/pdf': DocumentTypeCode.Pdf,
-		pdf: DocumentTypeCode.Pdf,
-		'application/msword': DocumentTypeCode.Word,
-		'application/vnd.openxmlformats-officedocument.wordprocessingml.document': DocumentTypeCode.Word,
-		doc: DocumentTypeCode.Word,
-		docx: DocumentTypeCode.Word,
-		'image/bmp': DocumentTypeCode.Image,
-		'image/gif': DocumentTypeCode.Image,
-		'image/jpg': DocumentTypeCode.Image,
-		'image/jpeg': DocumentTypeCode.Image,
-		'image/png': DocumentTypeCode.Image,
-		'image/tiff': DocumentTypeCode.Image,
-		bmp: DocumentTypeCode.Image,
-		gif: DocumentTypeCode.Image,
-		jpg: DocumentTypeCode.Image,
-		jpeg: DocumentTypeCode.Image,
-		png: DocumentTypeCode.Image,
-		tiff: DocumentTypeCode.Image,
-	};
-
-	public static getFileDocumentType(file: File): DocumentTypeCode;
-	public static getFileDocumentType(fileMimeType: string): DocumentTypeCode;
-	public static getFileDocumentType(file: File | string): DocumentTypeCode | null {
-		if (typeof file === 'string') return this._MIME_TYPE_DOCUMENT_TYPE_MAP[file];
-		else if (file?.type) return this._MIME_TYPE_DOCUMENT_TYPE_MAP[file.type];
-		return null;
-	}
-
-	public static getFileIcon(file: File): IconType;
-	public static getFileIcon(fileType: string): IconType;
-	public static getFileIcon(documentType: DocumentTypeCode): IconType;
-	public static getFileIcon(fileInfo: File | string | DocumentTypeCode): IconType {
-		let fileType: DocumentTypeCode;
-		if (typeof fileInfo === 'string' && Object.values(DocumentTypeCode).includes(fileInfo as DocumentTypeCode)) {
-			fileType = fileInfo as DocumentTypeCode;
-		} else {
-			fileType = FileUploadHelper.getFileDocumentType(typeof fileInfo === 'string' ? fileInfo : fileInfo?.type);
-		}
-
-		let fileIcon = { icon: 'insert_drive_file', label: 'File' };
-		if (fileType) fileIcon = this._FILE_ICONS[fileType];
-
-		return fileIcon;
-	}
-}
+import { DocumentTypeCode, FileUploadHelper, IconType } from './file-upload-helper';
 
 @Component({
 	selector: 'app-file-upload',
