+++ src/Spd.Presentation.Screening/ClientApp/src/app/shared/components/manual-submission-common.component.ts
@@ -35,7 +35,6 @@ import { CrrpRoutes } from 'src/app/modules/crrp-portal/crrp-routes';
 import { PssoRoutes } from 'src/app/modules/psso-portal/psso-routes';
 import { Address, AddressAutocompleteComponent } from 'src/app/shared/components/address-autocomplete.component';
 import { DialogComponent, DialogOptions } from 'src/app/shared/components/dialog.component';
-import { FileUploadComponent } from 'src/app/shared/components/file-upload.component';
 import { FormErrorStateMatcher } from 'src/app/shared/directives/form-error-state-matcher.directive';
 import { FormatDatePipe } from '../pipes/format-date.pipe';
 
@@ -493,7 +492,11 @@ export interface ManualSubmissionBody {
 								</div>
 								<div class="row my-4">
 									<div class="col-xxl-8 col-xl-10 col-lg-12">
-										<app-file-upload [maxNumberOfFiles]="1"></app-file-upload>
+										<app-file-upload
+											[maxNumberOfFiles]="1"
+											[control]="attachments"
+											[files]="attachments.value"
+										></app-file-upload>
 										<mat-error
 											class="mat-option-error"
 											*ngIf="
@@ -623,8 +626,6 @@ export class ManualSubmissionCommonComponent implements OnInit {
 	@Input() portal: PortalTypeCode | null = null;
 	@Input() isPsaUser: boolean | undefined = undefined;
 
-	@ViewChild(FileUploadComponent) fileUploadComponent!: FileUploadComponent;
-
 	constructor(
 		private router: Router,
 		private formBuilder: FormBuilder,
@@ -699,14 +700,6 @@ export class ManualSubmissionCommonComponent implements OnInit {
 	}
 
 	onSubmit() {
-		if (this.portal == PortalTypeCode.Crrp) {
-			const attachments =
-				this.fileUploadComponent.files && this.fileUploadComponent.files.length > 0
-					? this.fileUploadComponent.files[0]
-					: '';
-			this.form.controls['attachments'].setValue(attachments);
-		}
-
 		this.form.markAllAsTouched();
 
 		if (this.previousNameFlag.value != BooleanTypeCode.Yes) {
@@ -744,7 +737,7 @@ export class ManualSubmissionCommonComponent implements OnInit {
 			createRequest.requireDuplicateCheck = true;
 
 			const body: ManualSubmissionBody = {
-				ConsentFormFile: this.portal == PortalTypeCode.Crrp ? this.fileUploadComponent.files[0] : null,
+				ConsentFormFile: this.portal == PortalTypeCode.Crrp ? this.attachments.value : null,
 				ApplicationCreateRequestJson: JSON.stringify(createRequest),
 			};
 
@@ -1035,5 +1028,8 @@ export class ManualSubmissionCommonComponent implements OnInit {
 	}
 	get serviceTypeIsPssoVs(): boolean {
 		return this.form.get('serviceType')?.value === ServiceTypeCode.PssoVs;
+	}
+	get attachments(): FormControl {
+		return this.form.get('attachments') as FormControl;
 	}
 }
